requirements.txt
streamlit
openai
requests
promptflow
promptflow-tools
promptflow-core
promptflow-devkit
promptflow-azure
promptflow-tracing
faiss-cpu
sentence_transformers